+++ torch_frame/datasets/__init__.py
@@ -16,6 +16,7 @@ from .data_frame_benchmark import DataFrameBenchmark
 from .mercari import Mercari
 from .amazon_fine_food_reviews import AmazonFineFoodReviews
 from .diamond_images import DiamondImages
+from .huggingface_dataset import HuggingFaceDatasetDict
 
 real_world_datasets = [
     'Titanic',
@@ -39,4 +40,8 @@ synthetic_datasets = [
     'FakeDataset',
 ]
 
-__all__ = real_world_datasets + synthetic_datasets
+other_datasets = [
+    'HuggingFaceDatasetDict',
+]
+
+__all__ = real_world_datasets + synthetic_datasets + other_datasets